Added ability to map PT response messages into BO own error messages.
I added file pt_message_mapper.json where you can specify which PT message you want to map into your own message. The content of the file is evaluated to javascript object, every property of the object is treated as a message pattern for regular expression. Below is example content of the file:

{
  "Description .* 5000 characters": "This is my own error message that will be sent to the user"
}

In this case PT message "Description may not be more than 5000 characters." will be changed to message "This is my own error message that will be sent to the user".